:bug: Fix model validator

+++ src/models/office_manager.py
@@ -10,7 +10,7 @@ class StockBalance(BaseModel):
     unit_id: int
     ingredient_name: str
     days_left: int
-    stocks_count: int | float
+    stocks_count: float | int
     stocks_unit: str
